+++ lib/nosedrum/text_command/text_command.ex
